Struct primes::PrimeSet

pub struct PrimeSet {
    // some fields omitted
}

A prime generator, using the Sieve of Eratosthenes.

Create with let mut pset = PrimeSet::new(), and then use pset.iter() to iterate over all primes.

Methods

impl PrimeSet

fn new() -> PrimeSet

A new prime generator, primed with 2 and 3

fn expand(&mut self)

Finds one more prime, and adds it to the list

fn len(&self) -> usize

Number of primes found so far

fn list<'a>(&'a self) -> &'a [u64]

Return all primes found so far as a slice

fn generator<'a>(&'a mut self) -> PrimeSetIter<'a>

Iterator over all primes not yet found

fn iter<'a>(&'a mut self) -> PrimeSetIter<'a>

Iterator over all primes, starting with 2. If you don't care about the "state" of the PrimeSet, this is what you want!

fn iter_vec<'a>(&'a self) -> Iter<'a, u64>

Iterator over just the primes found so far

fn find(&mut self, n: u64) -> (usize, u64)

Find the next largest prime from a number Returns (idx, prime) Note that if n is prime, then the output will be (idx, n)

fn is_prime(&mut self, n: u64) -> bool

Check if a number is prime Note that this only requires primes up to n.sqrt() to be generated, and will generate them as necessary on its own.

fn find_vec(&self, n: u64) -> Option<(usize, u64)>

Find the next largest prime from a number, if it is within the already-found list Returns (idx, prime) Note that if n is prime, then the output will be (idx, n)

fn get(&mut self, index: usize) -> u64

Get the nth prime, even if we haven't yet found it

fn prime_factors(&mut self, n: u64) -> Vec<u64>

Get the prime factors of a number, starting from 2, including repeats