Struct primes::PrimeSet

pub struct PrimeSet { /* fields omitted */ }

A prime generator, using the Sieve of Eratosthenes.

Create with let mut pset = PrimeSet::new(), and then use pset.iter() to iterate over all primes.

Methods

impl PrimeSet

pub fn new() -> PrimeSet

A new prime generator, primed with 2 and 3

pub fn expand(&mut self)

Finds one more prime, and adds it to the list

pub fn len(&self) -> usize

Number of primes found so far

pub fn is_empty(&self) -> bool

pub fn list(&self) -> &[u64]

Return all primes found so far as a slice

Important traits for PrimeSetIter<'a>
impl<'a> Iterator for PrimeSetIter<'a> type Item = u64;

pub fn generator(&mut self) -> PrimeSetIter

Iterator over all primes not yet found

Important traits for PrimeSetIter<'a>
impl<'a> Iterator for PrimeSetIter<'a> type Item = u64;

pub fn iter(&mut self) -> PrimeSetIter

Iterator over all primes, starting with 2. If you don't care about the "state" of the PrimeSet, this is what you want!

pub fn iter_vec(&self) -> Iter<u64>

Iterator over just the primes found so far

pub fn find(&mut self, n: u64) -> (usize, u64)

Find the next largest prime from a number

Returns (idx, prime)

Note that if n is prime, then the output will be (idx, n)

pub fn is_prime(&mut self, n: u64) -> bool

Check if a number is prime

Note that this only requires primes up to n.sqrt() to be generated, and will generate them as necessary on its own.

pub fn find_vec(&self, n: u64) -> Option<(usize, u64)>

Find the next largest prime from a number, if it is within the already-found list

Returns (idx, prime)

Note that if n is prime, then the output will be (idx, n)

pub fn get(&mut self, index: usize) -> u64

Get the nth prime, even if we haven't yet found it

pub fn prime_factors(&mut self, n: u64) -> Vec<u64>

Get the prime factors of a number, starting from 2, including repeats